Fixed #374. Adjustements for new compilers. [r3674]

+++ solvers/optical/slab/python/fourier2d-python.cpp
@@ -22,7 +22,7 @@ static void FourierSolver2D_setMirrors(FourierSolver2D& self, py::object value) 
             PyErr_Clear();
             try {
                 if (py::len(value) != 2) throw py::error_already_set();
-                self.mirrors.reset(std::make_pair<double,double>(py::extract<double>(value[0]),py::extract<double>(value[1])));
+                self.mirrors.reset(std::make_pair<double,double>(double(py::extract<double>(value[0])), double(py::extract<double>(value[1]))));
             } catch (py::error_already_set) {
                 throw ValueError("None, float, or tuple of two floats required");
             }
